Added disabling user information to TPD
Compiler can be instructed to not generate the user and time
information in the header of the generated .cc and .hh files
by providing '-D' flag. However, ttcn3_makefilegen cannot
generate Makefile with this flag added to COMPILER_FLAGS.

To avoid Makefile postprocessing phase, a TPD schema of
'MakefileSettings' element has been extended with optional boolean
element 'disableUserInformation'. If set to true, '-D' flag
will be added to COMPILER_FLAGS in Makefile.

Signed-off-by: Jacek Klimkowicz <jakl@semihalf.com>