PresenceConstraint: generate code for the typed values validators
General idea is to extend the typed value descriptors (TTCN_Typedescriptor_t) with
the type specific validator function.

This validator is called at the begining of 'XXX_encode()' function and
at the end of 'XXX_decode()'.

The core of this validator is generated from the recursive analisys of
the type's sub-type asn-constraints.

Currently the main attention was for the constraint chains at the end of
which there is a NamedConstraint with the valid 'Presence' value
('ABSENT' and 'PRESENT').

Tested with the EtsiTs103097Module, IEEE1609dot2 and IEEE1609dot2BaseTypes ASN1 schemas.

In these shemas constraint chains contains the following constraint types:

 ContainedSubtypeConstraint
 SetOperationConstraint
 SingleInnerTypeConstraint
 NamedConstraint
 MultipleInnerTypeConstraint
 UnparsedMultipleInnerTypeConstraint

Before to start the implementaiton of the test cases,
I would like to know whether this proposal could be usefull to the project,
and whether it in the project's spirit.

Change-Id: Ied81ce1826de37226963138ce7dc9f5a6b231a53